Conflicts:
	src/backup.c
	src/btree.c
	src/btreeInt.h
	src/pager.c
	src/pcache1.c
	src/wal.c

+++ src/test2.c
@@ -578,7 +578,7 @@ static int fake_big_file(
   return TCL_ERROR;
 
   pVfs = sqlite3_vfs_find(0);
-  nFile = strlen(argv[2]);
+  nFile = (int)strlen(argv[2]);
   zFile = sqlite3_malloc( nFile+2 );
   if( zFile==0 ) return TCL_ERROR;
   memcpy(zFile, argv[2], nFile+1);
